requirements.txt
sacrebleu
sacremoses
scikit-learn
scipy
streamlit
titlecase
torch
tqdm
transformers==4.9.2
wandb
numpy
psutil
python-Levenshtein
pytorch-lightning==1.3.8
torchmetrics==0.6.2
PyYAML
regex
bokeh
dataclasses
colorlog
einops
fuzzywuzzy
ipywidgets   
joblib      
jsonschema
lexical-diversity